+++ Code/recommenderapp/prediction_scripts/search.py
@@ -23,7 +23,7 @@ project_dir = os.path.dirname(code_dir)
 
 class Search:
 
-    df = pd.read_csv(project_dir + "/Code/recommenderapp/prediction_scripts/data/movies.csv")
+    df = pd.read_csv(project_dir + "/recommenderapp/prediction_scripts/data/movies.csv")
 
     def __init__(self):
         pass
