--- test/README.md
@@ -1,1 +1,0 @@
-# Readme for test
